fix(platform): unbreak feature-switch lint and browser tests
- Restore the pinned web-backend transport for feature switches:
  routing through zeroClient$ created a static import cycle
  (feature-switch -> api-client -> fetch -> feature-switch),
  even though the runtime read is now sync from local storage.
- Drop the now-unused async on setupGlobalMethod$ to satisfy
  require-await; the awaited featureSwitches setup was removed
  in the parent commit.
- Switch test assertions to toBeTruthy/toBeFalsy per
  vitest/prefer-to-be-truthy.
- Move setMockFeatureSwitches into api-feature-switches.helpers.ts
  so msw/node does not leak into mocks/browser.ts via
  handlers/index.ts; this caused vitest browser tests to fail
  on Module "node:http" externalized for browser compatibility.
