+++ turbo/apps/platform/src/mocks/handlers/api-feature-switches.ts
@@ -3,16 +3,16 @@
  *
  * Mock handlers for /api/zero/feature-switches endpoint.
  *
- * Stateless: defaults return empty switches. Use `setMockFeatureSwitches` from
- * a test to override the GET response — it installs a fresh handler via
- * `server.use`, which is automatically reset between tests by the MSW
- * `server.resetHandlers()` afterEach hook.
+ * Stateless: defaults return empty switches. Tests override the GET response
+ * via `setMockFeatureSwitches` from `./api-feature-switches.helpers.ts` —
+ * that file imports `server` (msw/node) and is intentionally separate so
+ * browser tests, which transitively import `handlers/index.ts` to wire up
+ * `mocks/browser.ts`, do not pull in `node:http` through this module.
  */
 
 import { zeroFeatureSwitchesContract } from "@vm0/api-contracts/contracts/zero-feature-switches";
 
 import { mockApi } from "../msw-contract.ts";
-import { server } from "../server.ts";
 
 export const apiFeatureSwitchesHandlers = [
   mockApi(zeroFeatureSwitchesContract.get, ({ respond }) => {
@@ -27,19 +27,3 @@ export const apiFeatureSwitchesHandlers = [
     return respond(200, { deleted: true as const });
   }),
 ];
-
-export function setMockFeatureSwitches(
-  switches: Partial<Record<string, boolean>>,
-): void {
-  const sanitized: Record<string, boolean> = {};
-  for (const [key, value] of Object.entries(switches)) {
-    if (value !== undefined) {
-      sanitized[key] = value;
-    }
-  }
-  server.use(
-    mockApi(zeroFeatureSwitchesContract.get, ({ respond }) => {
-      return respond(200, { switches: sanitized });
-    }),
-  );
-}
