+++ turbo/apps/platform/src/views/lab-page/__tests__/lab-page.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 import { screen, waitFor } from "@testing-library/react";
 import { FeatureSwitchKey } from "@vm0/connectors/feature-switch-key";
-import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches";
+import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.helpers";
 import { testContext } from "../../../signals/__tests__/test-helpers.ts";
 import { detachedSetupPage } from "../../../__tests__/page-helper.ts";
