+++ turbo/apps/platform/src/views/zero-page/__tests__/auto-read-toggle.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 import { screen, waitFor } from "@testing-library/react";
 import { testContext } from "../../../signals/__tests__/test-helpers.ts";
 import { detachedSetupPage, click } from "../../../__tests__/page-helper.ts";
-import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.ts";
+import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.helpers.ts";
 import { mockSubagentThread } from "./chat-test-helpers.ts";
 import { autoReadEnabled$ } from "../../../signals/voice-io/voice-io-settings.ts";
