+++ turbo/apps/platform/src/views/zero-page/__tests__/chat-composer-model-selection-send.test.tsx
@@ -32,7 +32,7 @@ import {
   setMockOrgModelProviders,
   resetMockOrgModelProviders,
 } from "../../../mocks/handlers/api-org-model-providers.ts";
-import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.ts";
+import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.helpers.ts";
 import {
   mockChatLifecycle,
   sendMessageInUI,
