+++ turbo/apps/platform/src/views/zero-page/__tests__/model-provider-picker-display.test.tsx
@@ -26,7 +26,7 @@ import {
   setMockOrgModelProviders,
   resetMockOrgModelProviders,
 } from "../../../mocks/handlers/api-org-model-providers.ts";
-import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.ts";
+import { setMockFeatureSwitches } from "../../../mocks/handlers/api-feature-switches.helpers.ts";
 
 const context = testContext();
 const mockApi = createMockApi(context);
